requirements.txt

edge-tts
moviepy
Pillow
google-auth-httplib2
fastapi==0.115.2
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.9.2
email-validator==2.2.0
requests==2.32.3
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.154.0
pytz==2025.2
PyJWT




